requirements.txt
django==4.2.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
djoser==2.2.3
mysqlclient==2.2.4